Move Waypoint to schedule.py
