--- src/virtualship/expedition/waypoint.py
@@ -1,25 +1,0 @@
-"""Waypoint class."""
-
-from __future__ import annotations
-
-from datetime import datetime
-
-import pydantic
-
-from ..location import Location
-from .ship_config import InstrumentType
-
-
-class Waypoint(pydantic.BaseModel):
-    """A Waypoint to sail to with an optional time and an optional instrument."""
-
-    location: Location
-    time: datetime | None = None
-    instrument: InstrumentType | list[InstrumentType] | None = None
-
-    @pydantic.field_serializer("instrument")
-    def serialize_instrument(self, instrument):
-        """Ensure InstrumentType is serialized as a string (or list of strings)."""
-        if isinstance(instrument, list):
-            return [inst.value for inst in instrument]
-        return instrument.value if instrument else None
