+++ tests/expedition/test_schedule.py
@@ -5,9 +5,8 @@ import pyproj
 import pytest
 
 from virtualship import Location
-from virtualship.expedition import Waypoint
 from virtualship.expedition.do_expedition import _load_input_data
-from virtualship.expedition.schedule import Schedule, ScheduleError
+from virtualship.expedition.schedule import Schedule, ScheduleError, Waypoint
 from virtualship.utils import _get_ship_config
 
 projection = pyproj.Geod(ellps="WGS84")
